+++ src/test/java/org/opensearch/knn/memoryoptsearch/FaissMemoryOptimizedSearcherTests.java
@@ -40,6 +40,7 @@ import org.opensearch.knn.index.mapper.KNNVectorFieldMapper;
 import org.opensearch.knn.index.quantizationservice.QuantizationService;
 import org.opensearch.knn.index.query.FilterIdsSelector;
 import org.opensearch.knn.index.query.KNNQueryResult;
+import org.opensearch.knn.index.query.MemoryOptimizedSearchScoreConverter;
 import org.opensearch.knn.index.store.IndexInputWithBuffer;
 import org.opensearch.knn.index.store.IndexOutputWithBuffer;
 import org.opensearch.knn.index.vectorvalues.KNNByteVectorValues;
@@ -65,6 +66,7 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import static java.util.stream.Collectors.toMap;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.opensearch.knn.common.KNNConstants.ADC_ENABLED_FAISS_INDEX_INTERNAL_PARAMETER;
@@ -182,8 +184,8 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
         final TestingSpec testingSpec = new TestingSpec(
             VectorDataType.FLOAT,
             FLOAT_HNSW_INDEX_DESCRIPTION,
-            -1000000,
-            1000000,
+            -10,
+            10,
             FLOAT32_ENCODER_PARAMETERS
         );
 
@@ -226,8 +228,8 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
         final TestingSpec testingSpec = new TestingSpec(
             VectorDataType.FLOAT,
             FLOAT16_HNSW_INDEX_DESCRIPTION,
-            -65504,
-            65504,
+            -10,
+            10,
             FLOAT16_ENCODER_PARAMETERS
         );
 
@@ -244,7 +246,7 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
         doSearchTest(testingSpec, IndexingType.DENSE_NESTED);
     }
 
-    public void testADCWithBinaryQuantization() {
+    private void doTestADCWithBinaryQuantization(final SpaceType spaceType) {
         final TestingSpec adcEnabledSpec = new TestingSpec(
             VectorDataType.BINARY,
             BINARY_HSNW_INDEX_DESCRIPTION,
@@ -258,37 +260,45 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
             .build();
 
         adcEnabledSpec.isAdcEnabled = true;
-        // Define parameter options
-        List<IndexingType> indexingTypes = Arrays.asList(
-            IndexingType.DENSE,
-            IndexingType.DENSE_NESTED,
-            IndexingType.SPARSE,
-            IndexingType.SPARSE_NESTED
-        );
 
-        List<SpaceType> spaceTypes = Arrays.asList(SpaceType.L2, SpaceType.INNER_PRODUCT, SpaceType.COSINESIMIL);
+        doSearchTest(adcEnabledSpec, IndexingType.DENSE, spaceType, true, 0.8f);
+        doSearchTest(adcEnabledSpec, IndexingType.DENSE, spaceType, true, NO_FILTERING);
+        doSearchTest(adcEnabledSpec, IndexingType.DENSE, spaceType, false, 0.8f);
+        doSearchTest(adcEnabledSpec, IndexingType.DENSE, spaceType, false, NO_FILTERING);
 
-        List<Boolean> booleanOptions = Arrays.asList(true, false);
+        doSearchTest(adcEnabledSpec, IndexingType.DENSE_NESTED, spaceType, true, 0.8f);
+        doSearchTest(adcEnabledSpec, IndexingType.DENSE_NESTED, spaceType, true, NO_FILTERING);
+        doSearchTest(adcEnabledSpec, IndexingType.DENSE_NESTED, spaceType, false, 0.8f);
+        doSearchTest(adcEnabledSpec, IndexingType.DENSE_NESTED, spaceType, false, NO_FILTERING);
 
-        List<Object> filterOptions = Arrays.asList(0.8f, NO_FILTERING);
+        doSearchTest(adcEnabledSpec, IndexingType.SPARSE, spaceType, true, 0.8f);
+        doSearchTest(adcEnabledSpec, IndexingType.SPARSE, spaceType, true, NO_FILTERING);
+        doSearchTest(adcEnabledSpec, IndexingType.SPARSE, spaceType, false, 0.8f);
+        doSearchTest(adcEnabledSpec, IndexingType.SPARSE, spaceType, false, NO_FILTERING);
 
-        // // Generate cartesian product and run tests
-        for (SpaceType spaceType : spaceTypes) {
-            for (IndexingType indexingType : indexingTypes) {
-                for (Boolean boolOption : booleanOptions) {
-                    for (Object filterOption : filterOptions) {
-                        doSearchTest(adcEnabledSpec, indexingType, spaceType, boolOption, (float) filterOption);
-                    }
-                }
-            }
-        }
+        doSearchTest(adcEnabledSpec, IndexingType.SPARSE_NESTED, spaceType, true, 0.8f);
+        doSearchTest(adcEnabledSpec, IndexingType.SPARSE_NESTED, spaceType, true, NO_FILTERING);
+        doSearchTest(adcEnabledSpec, IndexingType.SPARSE_NESTED, spaceType, false, 0.8f);
+        doSearchTest(adcEnabledSpec, IndexingType.SPARSE_NESTED, spaceType, false, NO_FILTERING);
+    }
+
+    public void testADCWithBinaryQuantizationL2() {
+        doTestADCWithBinaryQuantization(SpaceType.L2);
+    }
+
+    public void testADCWithBinaryQuantizationIP() {
+        doTestADCWithBinaryQuantization(SpaceType.INNER_PRODUCT);
+    }
+
+    public void testADCWithBinaryQuantizationCosine() {
+        doTestADCWithBinaryQuantization(SpaceType.COSINESIMIL);
     }
 
     @SneakyThrows
     private void doSearchTest(final TestingSpec testingSpec, final IndexingType indexingType) {
         final List<SpaceType> spaceTypes;
         if (testingSpec.dataType != VectorDataType.BINARY) {
-            spaceTypes = Arrays.asList(SpaceType.L2, SpaceType.INNER_PRODUCT);
+            spaceTypes = Arrays.asList(SpaceType.L2, SpaceType.INNER_PRODUCT, SpaceType.COSINESIMIL);
         } else {
             spaceTypes = Arrays.asList(SpaceType.HAMMING);
         }
@@ -363,7 +373,12 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
 
         if (testingSpec.dataType == VectorDataType.FLOAT || testingSpec.dataType == VectorDataType.BYTE) {
             if (testingSpec.dataType == VectorDataType.FLOAT) {
-                queryForVectorReader = query = generateOneSingleFloatVector(DIMENSIONS, testingSpec.minValue, testingSpec.maxValue);
+                queryForVectorReader = query = generateOneSingleFloatVector(
+                    DIMENSIONS,
+                    testingSpec.minValue,
+                    testingSpec.maxValue,
+                    spaceType == SpaceType.COSINESIMIL
+                );
             } else if (testingSpec.dataType == VectorDataType.BYTE) {
                 queryForVectorReader = byteQuery = generateOneSingleByteVector(DIMENSIONS, testingSpec.minValue, testingSpec.maxValue);
                 query = convertToFloatArray(byteQuery);
@@ -380,7 +395,12 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
                 parentIds
             );
         } else if (testingSpec.isAdcEnabled) {
-            float[] rawFloat = (float[]) generateOneSingleFloatVector(DIMENSIONS, testingSpec.minValue, testingSpec.maxValue);
+            float[] rawFloat = generateOneSingleFloatVector(
+                DIMENSIONS,
+                testingSpec.minValue,
+                testingSpec.maxValue,
+                spaceType == SpaceType.COSINESIMIL
+            );
 
             (QuantizationService.getInstance()).transformWithADC(testingSpec.quantizationState, rawFloat, spaceType);
 
@@ -396,11 +416,15 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
                 FilterIdsSelector.FilterIdsSelectorType.BATCH.getValue(),
                 parentIds
             );
-
         } else if (testingSpec.dataType == VectorDataType.BINARY) {
             if (testingSpec.quantizationParams != null) {
-                float[] floatQuery = generateOneSingleFloatVector(DIMENSIONS, testingSpec.minValue, testingSpec.maxValue);
-                query = queryForVectorReader = (byte[]) QuantizationService.getInstance()
+                float[] floatQuery = generateOneSingleFloatVector(
+                    DIMENSIONS,
+                    testingSpec.minValue,
+                    testingSpec.maxValue,
+                    spaceType == SpaceType.COSINESIMIL
+                );
+                query = queryForVectorReader = QuantizationService.getInstance()
                     .quantize(
                         testingSpec.quantizationState,
                         floatQuery,
@@ -427,6 +451,14 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
 
         JNIService.free(indexPointer, KNNEngine.FAISS);
 
+        // Score transform in Faiss
+        for (int i = 0; i < resultsFromFaiss.length; i++) {
+            resultsFromFaiss[i] = new KNNQueryResult(
+                resultsFromFaiss[i].getId(),
+                KNNEngine.FAISS.score(resultsFromFaiss[i].getScore(), spaceType)
+            );
+        }
+
         // Search via VectorReader
         final KNNQueryResult[] resultsFromVectorReader = doSearchViaVectorReader(
             buildInfo,
@@ -434,7 +466,8 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
             testingSpec.isAdcEnabled ? VectorDataType.FLOAT : testingSpec.dataType,
             filteredIds,
             k,
-            doExhaustiveSearch
+            doExhaustiveSearch,
+            spaceType
         );
 
         // Validate results
@@ -467,7 +500,8 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
         VectorDataType vectorDataType,
         long[] filteredIds,
         final int k,
-        final boolean exhaustiveSearch
+        final boolean exhaustiveSearch,
+        final SpaceType spaceType
     ) {
         // Make KNN vector field info
         KNNCodecTestUtil.FieldInfoBuilder fieldInfoBuilder = KNNCodecTestUtil.FieldInfoBuilder.builder(TARGET_FIELD)
@@ -534,6 +568,9 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
         // Make results
         final TopDocs topDocs = knnCollector.topDocs();
         final ScoreDoc[] scoreDocs = topDocs.scoreDocs;
+        if (spaceType == SpaceType.COSINESIMIL) {
+            MemoryOptimizedSearchScoreConverter.convertToCosineScore(scoreDocs);
+        }
         assertTrue(scoreDocs.length >= k);
         final List<KNNQueryResult> results = new ArrayList<>();
         for (int i = 0; i < k; ++i) {
@@ -603,7 +640,8 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
                         buildInfo.documentIds,
                         DIMENSIONS,
                         testingSpec.minValue,
-                        testingSpec.maxValue
+                        testingSpec.maxValue,
+                        spaceType == SpaceType.COSINESIMIL
                     );
                     buildInfo.vectors = new SearchTestHelper.Vectors(floatVectors);
                     final KNNFloatVectorValues floatVectorValues = createKNNFloatVectorValues(documentIds, floatVectors);
@@ -614,7 +652,8 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
                         buildInfo.documentIds,
                         DIMENSIONS,
                         testingSpec.minValue,
-                        testingSpec.maxValue
+                        testingSpec.maxValue,
+                        false
                     );
                     assert (testingSpec.quantizationParams != null);
 
@@ -701,6 +740,21 @@ public class FaissMemoryOptimizedSearcherTests extends KNNTestCase {
             assertFalse(matchRatio < 0.8 && recall < 0.8);
         }
 
+        // Validate score values are the same
+        final Map<Integer, Float> faissIdScores = Arrays.stream(resultsFromFaiss)
+            .collect(toMap(KNNQueryResult::getId, KNNQueryResult::getScore));
+        final boolean isRunningInWindows = System.getProperty("os.name").toLowerCase().contains("win");
+        if (isRunningInWindows == false) {
+            // For unknown reason, this assertion is only failing in Windows
+            // Until root causing the issue, blocking assertion for Windows.
+            for (final KNNQueryResult result : resultsFromVectorReader) {
+                if (faissIdScores.containsKey(result.getId())) {
+                    final float scoreFromReader = result.getScore();
+                    final float faissScore = faissIdScores.get(result.getId());
+                    assertEquals(faissScore, scoreFromReader, 1e-3);
+                }
+            }
+        }
     }
 
     @SneakyThrows